Moves core lease errors from separate locations into new module where they can be considered together.
